Moved Responses to own folder. Made an abstract BaseResponse
