+++ espresso-server/app/src/androidTest/java/io/appium/espressoserver/lib/Handlers/Click.java
@@ -5,7 +5,7 @@ import android.support.test.espresso.ViewInteraction;
 import java.util.Map;
 
 import fi.iki.elonen.NanoHTTPD;
-import io.appium.espressoserver.lib.Http.AppiumResponse;
+import io.appium.espressoserver.lib.Http.Response.AppiumResponse;
 import io.appium.espressoserver.lib.Model.Element;
 
 import static android.support.test.espresso.action.ViewActions.click;
