+++ espresso-server/app/src/androidTest/java/io/appium/espressoserver/lib/Handlers/RequestHandler.java
@@ -3,7 +3,7 @@ package io.appium.espressoserver.lib.Handlers;
 import java.util.Map;
 
 import fi.iki.elonen.NanoHTTPD.IHTTPSession;
-import io.appium.espressoserver.lib.Http.AppiumResponse;
+import io.appium.espressoserver.lib.Http.Response.AppiumResponse;
 
 public interface RequestHandler {
     public AppiumResponse handle(IHTTPSession session, Map<String, String> uriParams);
