--- espresso-server/app/src/androidTest/java/io/appium/espressoserver/lib/Http/AppiumResponse.java
@@ -1,28 +1,0 @@
-package io.appium.espressoserver.lib.Http;
-
-import fi.iki.elonen.NanoHTTPD.Response;
-
-public class AppiumResponse {
-    private Response.Status status;
-    private Object response;
-
-    public AppiumResponse() {
-        status = Response.Status.OK;
-    }
-
-    public Object getResponse() {
-        return response;
-    }
-
-    public void setResponse(Object response) {
-        this.response = response;
-    }
-
-    public Response.Status getStatus() {
-        return status;
-    }
-
-    public void setStatus(Response.Status status) {
-        this.status = status;
-    }
-}
