+++ espresso-server/app/src/androidTest/java/io/appium/espressoserver/lib/Http/Response/AppiumResponse.java
@@ -1,0 +1,6 @@
+package io.appium.espressoserver.lib.Http.Response;
+
+import fi.iki.elonen.NanoHTTPD.Response;
+
+public class AppiumResponse extends BaseResponse {
+}
