+++ espresso-server/app/src/androidTest/java/io/appium/espressoserver/lib/Http/Response/NotFoundResponse.java
@@ -1,4 +1,4 @@
-package io.appium.espressoserver.lib.Http;
+package io.appium.espressoserver.lib.Http.Response;
 
 import fi.iki.elonen.NanoHTTPD;
