+++ espresso-server/app/src/androidTest/java/io/appium/espressoserver/lib/Http/Router.java
@@ -10,6 +10,8 @@ import io.appium.espressoserver.lib.Handlers.Click;
 import io.appium.espressoserver.lib.Handlers.Finder;
 import io.appium.espressoserver.lib.Handlers.RequestHandler;
 import io.appium.espressoserver.lib.Handlers.CreateSession;
+import io.appium.espressoserver.lib.Http.Response.AppiumResponse;
+import io.appium.espressoserver.lib.Http.Response.NotFoundResponse;
 
 
 public class Router {
